% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_participants}
\alias{get_participants}
\title{Get participant information}
\usage{
get_participants(df, c)
}
\arguments{
\item{df}{Data frame with 'p' and 'condition' columns}

\item{c}{String used to filter() 'condition' column}
}
\value{
list('participants' = p, 'n' = count(p))
}
\description{
Get participant information
}
