% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfactory.R
\name{process_ant}
\alias{process_ant}
\title{Process meditation log data file}
\usage{
process_ant(ant_file, p, time)
}
\arguments{
\item{ant_file}{Path to log file}

\item{p}{Participant number}

\item{time}{Time-point at which participant completed ANT}

\item{ant_file}{Path to ANT file}

\item{p}{Participant number}
}
\value{
Data frame
Process expfactory Attention Network Test data file

\code{process_ant()} processes data files generated by the The Experiment Factory
implementation of the Attention Network Task
(\url{https://expfactory.github.io/attention_network_task.html}).

Data frame
}
\description{
\code{process_meditation_log()} processes meditation log files
recording how much practice a participant has done
}
\keyword{ANT}
\keyword{expfactory}
\keyword{meditation}
