% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfactory.R
\name{process_new_expfactory_survey}
\alias{process_new_expfactory_survey}
\title{Process expfactory 3.X survey data}
\usage{
process_new_expfactory_survey(token, survey, flat = FALSE)
}
\arguments{
\item{token}{token number}

\item{survey}{path to JSON survey file}

\item{unflatten=FALSE}{Unflatten "flat" JSON}
}
\value{
Data frame (long format)
}
\description{
Process expfactory 3.X survey data.
}
\keyword{expfactory}
