% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfactory.R
\name{process_surveys}
\alias{process_surveys}
\title{Process experiment factory surveys.}
\usage{
process_surveys(s, base_path)
}
\arguments{
\item{s}{list of survey names}

\item{base_path}{path to search}
}
\value{
Data frame
}
\description{
Process experiment factory surveys.
}
