% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfactory.R
\name{process_surveys3}
\alias{process_surveys3}
\title{Process expfactory survey data}
\usage{
process_surveys3(base_path)
}
\arguments{
\item{base_path}{data path}
}
\value{
data frame with [survey, value, participant, question]
}
\description{
Process expfactory survey data
}
\examples{
process_surveys3(base_path)
}
