% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breath_counting_accuracy.R
\name{summarise_breath_counting_accuracy}
\alias{summarise_breath_counting_accuracy}
\title{Summarise breath counting accuracy}
\usage{
summarise_breath_counting_accuracy(df)
}
\arguments{
\item{df}{Data frame returned by breath_counting_accuracy()}
}
\value{
data.frame
}
\description{
Summarise breath counting accuracy
}
\examples{
summarise_breath_counting_accuracy(df)
}
