% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{summarySE}
\alias{summarySE}
\title{Summarize data}
\usage{
summarySE(data = NULL, measurevar, groupvars = NULL, na.rm = FALSE,
  conf.interval = 0.95, .drop = TRUE)
}
\arguments{
\item{data}{Data frame}

\item{measurevar}{String containing name of a column that contains the variable to be summarized}

\item{groupvars}{Vector of strings naming columns that contain grouping variables}

\item{na.rm}{Boolean indicating whether to ignore NA's}

\item{conf.interval}{Number < 1 indicating percent range of the confidence interval (default=.95)}
}
\value{
Data frame
}
\description{
Calculates count, mean, standard deviation, standard error of the mean, and confidence interval.
}
\details{
See http://www.cookbook-r.com/Manipulating_data/Summarizing_data/
}
