% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfactory.R
\name{process_ant}
\alias{process_ant}
\title{Process expfactory Attention Network Test data file}
\usage{
process_ant(ant_file, p, time, json = TRUE)
}
\arguments{
\item{ant_file}{Path to ANT file}

\item{p}{Participant number}

\item{time}{Time-point at which participant completed ANT}
}
\value{
Data frame
}
\description{
\code{process_ant()} processes data files generated by the The Experiment Factory
implementation of the Attention Network Task
(\url{https://expfactory.github.io/attention_network_task.html}).
}
\keyword{ANT}
\keyword{expfactory}
