% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfactory.R
\name{process_expfactory_survey}
\alias{process_expfactory_survey}
\title{Process expfactory survey data}
\usage{
process_expfactory_survey(p, survey)
}
\arguments{
\item{p}{Participant Number}

\item{survey}{Survey name}
}
\value{
Data frame (long format)
# http://expfactory.readthedocs.io/en/latest/development.html#contributing-to-surveys
}
\description{
Process expfactory survey data
}
\keyword{expfactory}
